requirements.txt
discord.py
discord.py[voice]
asyncio
pip
youtube_dl
websockets
opuslib
PyNaCl
wikipedia-api

